Get latest tiger mosquito colonisation data in France from the online map

Get latest tiger mosquito colonisation data in France from the online map at https://signalement-moustique.anses.fr/signalement_albopictus/colonisees.

Usage

get_tiger_mosquito_colonisation_in_france(output = c("sf", "geojson"))

Arguments

output

The output format. Can be either "sf" or "geojson".

Value

If output = "sf", an sf object with the tiger mosquito colonisation data in France. All polygons that exist on the map are covering the territories where the tiger mosquito (Aedes Albopictus) has alreadt been detected in France. If output = "geojson", a GeoJSON in a character vector with the tiger mosquito colonisation data in France. Using the 'geojson' option allows to use the package without the sf package. You can therefore just save the resulting string to a text file. See examples.

Examples

# \donttest{
# get the latest data in `sf`
x <- get_tiger_mosquito_colonisation_in_france(output = "sf")

# save the data to GeoPackage file with `sf`
if(interactive()) {
library(sf)
st_write(x, "tiger_mosquito_colonisation_in_france.gpkg")
}
# }

# get the latest data in `geojson`
x <- get_tiger_mosquito_colonisation_in_france(output = "geojson")

# save the data to text file
if(interactive()) {
writeLines(x, "tiger_mosquito_colonisation_in_france.geojson")
}